reassemble and eukcc

+++ bin/metawrap-modules/reassemble_bins.sh
@@ -400,7 +400,10 @@ if [ "$run_checkm" = true ] || [ "$run_eukcc" = true ] ; then
 		error "there are no good bins found in ${out}/reassembled_best_bins - something went wrong with choosing the best bins between the reassemblies."
 	fi
 
+fi
 
+if [ "$run_checkm" = true ] && [ "$run_eukcc" = false ] ; then
+        ###############   Re-running CheckM on the best reasembled bins ###############
 	comm "Re-running CheckM on the best reasembled bins."
 	if [[ -d ${out}/reassembled_bins.checkm ]]; then rm -r ${out}/reassembled_bins.checkm; fi
 	mkdir ${out}/tmp
@@ -424,6 +427,26 @@ if [ "$run_checkm" = true ] || [ "$run_eukcc" = true ] ; then
 	grep orig ${out}/work_files/reassembled_bins.stats >> ${out}/original_bins.stats
 	${SOFT}/plot_reassembly.py $out $comp $cont ${out}/reassembled_bins.stats ${out}/original_bins.stats
 	if [[ $? -ne 0 ]]; then error "Something went wrong with plotting the reassembly summary plots. Exiting..."; fi
+
+elif [ "$run_eukcc" = true ] ; then
+	###############   Re-running EukCC on the best reasembled bins ###############
+	comm "Re-running EukCC on the best reasembled bins."
+        if [[ -d ${out}/reassembled_bins.eukcc ]]; then rm -r ${out}/reassembled_bins.eukcc; fi
+	eukcc folder --db $EUKCC2_DB --out ${out}/reassembled_bins.eukcc --threads $threads ${out}/reassembled_bins
+        ${SOFT}/bins_stats.py -i ${out}/reassembled_bins -o ${out}/reassembled_bins.eukcc/basic.csv
+        join -j 1 -t $'\t' <(cat ${out}/reassembled_bins.eukcc/eukcc.csv | (sed -u 1q; sort)) <(cat ${out}/reassembled_bins.eukcc/basic.csv | (sed -u 1q; sort)) > ${out}/reassembled_bins.eukcc/eukcc.tsv
+        if [[ ! -s ${out}/reassembled_bins.eukcc/eukcc.tsv ]]; then error "Something went wrong with running EukCC. Exiting..."; fi
+        comm "Finalizing EukCC stats..."
+        ${SOFT}/summarize_eukcc.py ${out}/reassembled_bins.eukcc/eukcc.tsv | (read -r; printf "%s\n" "$REPLY"; sort) > ${out}/reassembled_bins.stats
+        if [[ $? -ne 0 ]]; then error "Cannot make eukcc summary file. Exiting."; fi
+
+        comm "you will find the info on the final reassembled bins in ${out}/reassembled_bins.stats"
+
+        comm "making reassembly N50, compleiton, and contamination summary plots."
+        head -n 1 ${out}/work_files/reassembled_bins.stats > ${out}/original_bins.stats
+        grep orig ${out}/work_files/reassembled_bins.stats >> ${out}/original_bins.stats
+        ${SOFT}/plot_reassembly.py $out $comp $cont ${out}/reassembled_bins.stats ${out}/original_bins.stats
+        if [[ $? -ne 0 ]]; then error "Something went wrong with plotting the reassembly summary plots. Exiting..."; fi
 fi
 
 comm "you will find the final bins in ${out}/reassembled_bins"
